requirements.txt
streamlit
pandas
openpyxl
google-generativeai
xlsxwriter
Pillow
st-gsheets-connection
gspread